requirements.txt
black
ruff
pytest

django
djangorestframework
django-cors-headers

celery
redis

chromadb
pydantic

pdfplumber

python-decouple

requests